Fix weird mixing of layers from main and spare pages (Issue 259)

+++ global.h
@@ -338,7 +338,7 @@ GFX2_GLOBAL short Main_magnifier_offset_X;
 /// Y position (in image space) of the pixel to display in the top left corner of the magnified view.
 GFX2_GLOBAL short Main_magnifier_offset_Y;
 /// Index of layer currently being edited
-GFX2_GLOBAL int Main_current_layer;
+GFX2_GLOBAL byte Main_current_layer;
 /// Bitfield that records which layers are visible. 2^0 for 0, 2^1 for 1, 2^2 for 2, etc.
 GFX2_GLOBAL dword Main_layers_visible;
 
@@ -402,7 +402,7 @@ GFX2_GLOBAL short Spare_magnifier_offset_X;
 /// Y position (in image space) of the pixel to display in the top left corner of the magnified view.
 GFX2_GLOBAL short Spare_magnifier_offset_Y;
 /// Index of layer currently being edited
-GFX2_GLOBAL short Spare_current_layer;
+GFX2_GLOBAL byte Spare_current_layer;
 /// Bitfield that records which layers are visible. 2^0 for 0, 2^1 for 1, 2^2 for 2, etc.
 GFX2_GLOBAL dword Spare_layers_visible;
 // -- Image backups
